cut files for 1/16" and 1/8" thick Delrin sheet
cut files for 1/16" and 1/8" thick Delrin sheet with the parts arranged with fairly efficient material usage. These were arranged for sheet with dimension 12" x 24". Some parts contain counterbores to be laser etched in order to provide clearance for the fillet under the rivet head. Some parts, like the gripper finger, are cut in greater quantity that the others because they tend to be easier to break.